pyproject.toml
[project]
name = "big-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastexcel>=0.19.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pathlib>=1.0.1",
    "polars>=1.40.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "tabulate>=0.10.0",
    "tqdm>=4.67.3",
]
